backend/requirements.txt
fastapi
uvicorn
requests
pydantic
sqlalchemy
chromadb
python-multipart
groq
openai
python-docx
python-dotenv
pypdf
pytesseract
opencv-python
numpy
twilio
google-api-python-client
google-auth
google-auth-oauthlib
reportlab
